# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ISP Ops Platform API
  version: 0.1.0
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
paths:
...
  /incidents:
    get:
      summary: List operational incidents
      responses:
        "200":
          description: Incident list
    post:
      summary: Create operational incident
      security:
        - ApiKeyAuth: []
      responses:
        "201":
          description: Incident created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
  /incidents/{code}:
    get:
      summary: Incident detail
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Incident with events and impacts
        "404":
          description: Incident not found
  /incidents/{code}/events:
    post:
      summary: Add incident timeline event
      security:
        - ApiKeyAuth: []
      responses:
        "201":
          description: Incident event created
        "400":
          description: Invalid payload
        "404":
          description: Incident not found
...
  /documentation/documents:
    get:
      summary: List technical documents
      parameters:
        - name: objectType
          in: query
          schema:
            type: string
        - name: objectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Document list
    post:
      summary: Create technical document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectType, objectId, title, bodyMd]
              properties:
                objectType:
                  type: string
                objectId:
                  type: string
                title:
                  type: string
                bodyMd:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Document created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
  /documentation/evidence:
    get:
      summary: List evidence files
      parameters:
        - name: objectType
          in: query
          schema:
            type: string
        - name: objectId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Evidence list
    post:
      summary: Create evidence record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [objectType, objectId, filename, storageKey]
              properties:
                objectType:
                  type: string
                objectId:
                  type: string
                filename:
                  type: string
                storageKey:
                  type: string
                contentType:
                  type: string
                reason:
                  type: string
      responses:
        "201":
          description: Evidence record created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
...
  /changes:
    get:
      summary: List change requests
      responses:
        "200":
          description: Change request list
    post:
      summary: Create change request
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description]
              properties:
                title:
                  type: string
                description:
                  type: string
                riskLevel:
                  type: string
                plannedStart:
                  type: string
                plannedEnd:
                  type: string
                impacts:
                  type: array
      responses:
        "201":
          description: Change request created
        "400":
          description: Invalid payload
        "503":
          description: PostgreSQL required
  /changes/{id}/impacts:
    get:
      summary: List change impacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Impact list
        "503":
          description: PostgreSQL required
  /changes/{id}/approve:
    post:
      summary: Approve change request
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Change request approved
        "503":
          description: PostgreSQL required
...